Move kotlin files to the 'kotlin' folder
