--- app/src/main/java/com/github/versus/chats/MessageAdapter.kt
@@ -1,63 +1,0 @@
-package com.github.versus.chats
-
-import android.content.Context
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.github.versus.R
-import com.github.versus.user.UserAdapter
-import com.google.firebase.auth.FirebaseAuth
-
-class MessageAdapter(val context: Context, val messageList: ArrayList<Message>) : RecyclerView.Adapter<RecyclerView.ViewHolder>(){
-
-    val ITEM_RECEIVE = 1;
-    val ITEM_SENT = 2;
-
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        if(viewType == 1){
-            val view : View = LayoutInflater.from(context).inflate(R.layout.receive_layout, parent, false)
-            return ReceiveViewHolder(view)
-        }else{
-            val view : View = LayoutInflater.from(context).inflate(R.layout.send_layout, parent, false)
-            return SentViewHolder(view)
-        }
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        val currentMessage = messageList[position]
-        if(FirebaseAuth.getInstance().currentUser?.uid.equals(currentMessage.sender.uid)){
-            return ITEM_SENT
-        }
-        return ITEM_RECEIVE
-    }
-
-
-    override fun getItemCount(): Int {
-        return messageList.size
-    }
-
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-        val currentMessage = messageList[position]
-
-        if(holder.javaClass == SentViewHolder::class.java){
-            val viewHolder = holder as SentViewHolder
-             viewHolder.sentMessage.text = currentMessage.body
-
-
-        }else{
-            val viewHolder = holder as ReceiveViewHolder
-            viewHolder.receiveMessage.text = currentMessage.body
-        }
-    }
-
-    class SentViewHolder(itemView : View) : RecyclerView.ViewHolder(itemView){
-        val sentMessage = itemView.findViewById<TextView>(R.id.txt_sent_message)
-    }
-    class ReceiveViewHolder(itemView : View): RecyclerView.ViewHolder(itemView){
-        val receiveMessage = itemView.findViewById<TextView>(R.id.txt_receive_message)
-
-    }
-}
